flaot32

+++ model.py
@@ -137,7 +137,8 @@ class Summarizer(object):
                 pad = tf.zeros([config.HIDDEN_SIZE])
                 flags = tf.Variable(embed_init([3, config.HIDDEN_SIZE],
                                     dtype=tf.float32))
-                embeddings = tf.constant(data.load_embeddings(self.data_path))
+                embeddings = tf.constant(data.load_embeddings(self.data_path),
+                                         dtype=float32)
                 self.embeddings = tf.concat(1, [pad, flags, embeddings])
             else:
                 self.embeddings = tf.Variable(embed_init([self.enc_vocab,
